==> notebooks/4) Modell Evaluation.ipynb <==
import pandas as pd
import sys
sys.path.append("../src")
from RequestMapper import RequestMapper

from sklearn.metrics import mean_absolute_error
from sklearn.metrics import r2_score
from sklearn.metrics import accuracy_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import f1_score

pipelines = pd.read_pickle("../resource/Pipelines/ModelPipelines.pkl")


## Evaluation of the Age-Regressor 

pipeline_age = pipelines.values[0]

X_test = pipeline_age.transformation.X_test
X_test_numerical_transformed = pipeline_age.transformation.X_test_numerical_transformed
X_test_text_transformed = pipeline_age.transformation.X_test_text_transformed
y_test = pipeline_age.transformation.y_test

y_pred_num = pipeline_age.modelling.numerical_model.predict(X_test_numerical_transformed)
y_pred_text = pipeline_age.modelling.text_model.predict(X_test_text_transformed)
y_pred_stacking = pipeline_age.modelling.weighted_prediction(X_test, None, "regression")

def create_metric(y_pred_type, y_pred, y_test):
    print(f"Report for {y_pred_type} Dataset \n\n")
    print("Correlation matrix:")
    df_numerical_eval = pd.DataFrame({f"{y_pred_type} predictions":y_pred, "Actual y":y_test})
    print(df_numerical_eval.corr())
    print(f"\n\nMean absolute error: {mean_absolute_error(y_test, y_pred)}")
    print(f"R2-Score: {r2_score(y_test, y_pred)}")

create_metric("Numerical", y_pred_num, y_test)

create_metric("Text", y_pred_text, y_test)

create_metric("stacking", y_pred_stacking, y_test)

weights = pipeline_age.modelling.weights

print(f"The models will be weighted as follows:\n textual model: {weights[0]*100}% \n numerical model: {weights[1]*100}%")

### Evaluation of the Gender-Classifier 

pipeline_gender = pipelines.values[1]

X_test = pipeline_gender.transformation.X_test
X_test_numerical_transformed = pipeline_gender.transformation.X_test_numerical_transformed
X_test_text_transformed = pipeline_gender.transformation.X_test_text_transformed
y_test = pipeline_gender.transformation.y_test

y_pred_num = pipeline_gender.modelling.numerical_model.predict(X_test_numerical_transformed)
y_pred_text = pipeline_gender.modelling.text_model.predict(X_test_text_transformed)
y_pred_stacking = pipeline_gender.modelling.weighted_prediction(X_test, None)

def create_metric(y_pred_type, y_pred, y_test):
    print(f"Report for {y_pred_type} Dataset \n\n")
    print("Confusion matrix:")
    print(confusion_matrix(y_test, y_pred))
    print(f"\n\nAccuracy: {accuracy_score(y_test, y_pred)}")
    print(f"F1-Score: {f1_score(y_test, y_pred, average='macro')}")
    
    print("F1-Score = Accuracy im binären Fall")

create_metric("Numerical", y_pred_num, y_test)

create_metric("Text", y_pred_text, y_test)

create_metric("Weighted", y_pred_stacking, y_test)

weights = pipeline_gender.modelling.weights

print(f"The models will be weighted as follows:\n textual model: {weights[0]*100}% \n numerical model: {weights[1]*100}%")

## Evaluation of the Sign-Classifier 

pipeline_sign = pipelines.values[2]

X_test = pipeline_sign.transformation.X_test
X_test_numerical_transformed = pipeline_sign.transformation.X_test_numerical_transformed
X_test_text_transformed = pipeline_sign.transformation.X_test_text_transformed
y_test = pipeline_sign.transformation.y_test

y_pred_num = pipeline_sign.modelling.numerical_model.predict(X_test_numerical_transformed)

y_pred_text = pipeline_sign.modelling.text_model.predict(X_test_text_transformed)
y_pred_stacking = pipeline_sign.modelling.weighted_prediction(X_test, None)

def create_metric(y_pred_type, y_pred, y_test):
    print(f"Report for {y_pred_type} Dataset \n\n")
    
    print("Relative Share of of each classes:")
    print(y_test.value_counts()/y_test.size)
    print("The values above can be used to compare the model performance against with the most expected class")
    
    print(f"\n\nAccuracy: {accuracy_score(y_test, y_pred)}")
    print(f"F1-Score: {f1_score(y_test, y_pred, average='macro')}")

create_metric("Numerical", y_pred_num, y_test)

create_metric("Text", y_pred_text, y_test)

create_metric("Stacking", y_pred_stacking, y_test)

weights = pipeline_sign.modelling.weights

print(f"The models will be weighted as follows:\n textual model: {weights[0]*100}% \n numerical model: {weights[1]*100}%")

## Evaluation of the Topic-Classifier 

pipeline_topic = pipelines.values[3]

X_test = pipeline_topic.transformation.X_test
X_test_numerical_transformed = pipeline_topic.transformation.X_test_numerical_transformed
X_test_text_transformed = pipeline_topic.transformation.X_test_text_transformed
y_test = pipeline_topic.transformation.y_test

y_pred_num = pipeline_topic.modelling.numerical_model.predict(X_test_numerical_transformed)
y_pred_text = pipeline_topic.modelling.text_model.predict(X_test_text_transformed)
y_pred_stacking = pipeline_topic.modelling.weighted_prediction(X_test, None)

create_metric("Numerical", y_pred_num, y_test)

create_metric("Text", y_pred_text, y_test)

create_metric("Stacking", y_pred_stacking, y_test)

weights = pipeline_topic.modelling.weights

print(f"The models will be weighted as follows:\n textual model: {weights[0]}% \n numerical model: {weights[1]}%")
