==> notebooks/7) Request Mapper.ipynb <==
import pandas as pd
import sys
sys.path.append("../src")
from RequestMapper import RequestMapper

# initialisieren des Pipeline Mappers
df_full_preprocessed = pd.read_pickle("../resource/df_full_preprocessed.pkl")
pipelines_predictions = pd.read_pickle("../resource/Pipelines/ModelPipelines.pkl")
pipelines_cluster = pd.read_pickle("../resource/Pipelines/ClusteringPipeline.pkl")
pipelines_knn = pd.read_pickle("../resource/Pipelines/KNearestNeighborsPipeline.pkl")

rm = RequestMapper(df_full_preprocessed, pipelines_predictions, pipelines_cluster, pipelines_knn)


## Test Age

# Variable, die an die API übergeben werden

target_variable = "age"
age = None
sign = "Leo"
topic = "InvestmentBanking"
gender = "male"

text = "Jimmy, thank you!! I really like your videos. They're helping me a lot since I can't go to the library. But I loved live study with me too and I guess you've deleted them bcz I can't find them anymore. Are you gonna do more live streams? That would be really cool. It feels more real ;) Greetings from Europe!"

# verwendet nur text
print(rm.predict_text(text, target_variable))

# verwendet text und numerischen Daten
# wichtig ist hier das entsprechend der Zielvariable die variable age nicht übermittelt wird
print(rm.predict_numerical(target_variable, text, sign=sign, topic=topic, gender=gender))

# Nutz beide Modelle mittels einer optimierten Gewichtung und mischt die Vorhersagen.
# wichtig ist hier das entsprechend der Zielvariable die variable age nicht übermittelt wird
print(rm.predict_weighted(target_variable, text, sign=sign, topic=topic, gender=gender))

### Test Topic

# Variable, die an die API übergeben werden

target_variable = "topic"
age = 38
sign = "Leo"
topic = None#"InvestmentBanking"
gender = "male"

text = "Jimmy, thank you!! I really like your videos. They're helping me a lot since I can't go to the library. But I loved live study with me too and I guess you've deleted them bcz I can't find them anymore. Are you gonna do more live streams? That would be really cool. It feels more real ;) Greetings from Europe!"

# verwendet nur text
print(rm.predict_text(text, target_variable))

# verwendet text und numerischen Daten
# wichtig ist hier das entsprechend der Zielvariable die variable age nicht übermittelt wird
print(rm.predict_numerical(target_variable, text, age=age, sign=sign, topic=topic, gender=gender))

# Nutz beide Modelle mittels einer optimierten Gewichtung und mischt die Vorhersagen.
# wichtig ist hier das entsprechend der Zielvariable die variable age nicht übermittelt wird
print(rm.predict_weighted(target_variable, text, age=age, sign=sign, topic=topic, gender=gender))

### Test Clustering 

# Print cluster for text
print(rm.transform_cluster(mode="text", text=text))

# Print cluster for numerical 
print(rm.transform_cluster(mode="numerical", text=text, age=13, sign="Scorpio", gender="male", topic="InvestmentBanking"))

## Test Nearest Neighbors

# Print cluster for text
print(rm.transform_knn(mode="text", text=text))
print("\n\n")
# Print cluster for numerical 
print(rm.transform_knn(mode="numerical", text=text, age=13, sign="Scorpio", gender="male", topic="InvestmentBanking"))
